requirements.txt
bottle==0.12.7
Pillow==2.7.0
Shapely==1.3.0
geojson==1.0.9
ExifRead==1.4.2
